requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
distro==1.9.0
Django==5.1.6
et_xmlfile==2.0.0
exceptiongroup==1.2.2
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
lxml==5.3.0
MarkupSafe==3.0.2
olefile==0.47
openai==1.61.1
openpyxl==3.1.5
pillow==11.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyhwp==0.1b15
PyMuPDF==1.25.3
python-docx==1.1.2
python-pptx==1.0.2
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
XlsxWriter==3.2.2

pip~=25.0.1
setuptools~=75.6.0